Rename SRV target to Subdomain target
