--- subdomains/src/Filament/Admin/Resources/SrvTargets/Pages/ManageSrvTargets.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Boy132\Subdomains\Filament\Admin\Resources\SrvTargets\Pages;
-
-use Boy132\Subdomains\Filament\Admin\Resources\SrvTargets\SrvTargetResource;
-use Filament\Resources\Pages\ManageRecords;
-
-class ManageSrvTargets extends ManageRecords
-{
-    protected static string $resource = SrvTargetResource::class;
-}
