+++ subdomains/src/Filament/Admin/Resources/SubdomainTargets/Pages/ManageSubdomainTargets.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Boy132\Subdomains\Filament\Admin\Resources\SubdomainTargets\Pages;
+
+use Boy132\Subdomains\Filament\Admin\Resources\SubdomainTargets\SubdomainTargetResource;
+use Filament\Resources\Pages\ManageRecords;
+
+class ManageSubdomainTargets extends ManageRecords
+{
+    protected static string $resource = SubdomainTargetResource::class;
+}
