+++ subdomains/src/Filament/Admin/Resources/SubdomainTargets/SubdomainTargetResource.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace Boy132\Subdomains\Filament\Admin\Resources\SrvTargets;
+namespace Boy132\Subdomains\Filament\Admin\Resources\SubdomainTargets;
 
 use App\Filament\Admin\Resources\Nodes\Pages\EditNode;
 use App\Models\Node;
-use Boy132\Subdomains\Filament\Admin\Resources\SrvTargets\Pages\ManageSrvTargets;
+use Boy132\Subdomains\Filament\Admin\Resources\SubdomainTargets\Pages\ManageSubdomainTargets;
 use Filament\Resources\Resource;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Columns\TextInputColumn;
 use Filament\Tables\Table;
 
-class SrvTargetResource extends Resource
+class SubdomainTargetResource extends Resource
 {
     protected static ?string $model = Node::class;
 
-    protected static ?string $slug = 'srv-targets';
+    protected static ?string $slug = 'subdomain-targets';
 
     protected static string|\BackedEnum|null $navigationIcon = 'tabler-world-www';
 
     public static function getModelLabel(): string
     {
-        return trans('subdomains::strings.srv_target');
+        return trans('subdomains::strings.subdomain_target');
     }
 
     public static function getNavigationGroup(): ?string
@@ -37,12 +37,12 @@ class SrvTargetResource extends Resource
                     ->url(fn (Node $node) => user()?->can('update', $node) ? EditNode::getUrl(['record' => $node]) : null),
                 TextColumn::make('fqdn')
                     ->label(trans('admin/node.table.address')),
-                TextInputColumn::make('srv_target')
-                    ->label(trans('subdomains::strings.srv_target'))
-                    ->placeholder(trans('subdomains::strings.no_srv_target'))
+                TextInputColumn::make('subdomain_target')
+                    ->label(trans('subdomains::strings.subdomain_target'))
+                    ->placeholder(trans('subdomains::strings.no_subdomain_target'))
                     ->updateStateUsing(function (Node $node, $state) {
                         $node->forceFill([
-                            'srv_target' => $state,
+                            'subdomain_target' => $state,
                         ])->save();
                     }),
             ])
@@ -54,7 +54,7 @@ class SrvTargetResource extends Resource
     public static function getPages(): array
     {
         return [
-            'index' => ManageSrvTargets::route('/'),
+            'index' => ManageSubdomainTargets::route('/'),
         ];
     }
 }
